requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
frozenlist==1.4.1
idna==3.7
jalali_core==1.0.0
jdatetime==5.0.0
multidict==6.0.5
peewee==3.17.5
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-magic==0.4.27
TgCrypto==1.2.5
yarl==1.9.4